chore: rename namespaces
