+++ Conflux.Domain.Logic.Tests/Services/UserSessionServiceTests.cs
@@ -8,7 +8,7 @@ using System.Text.Json;
 using Conflux.Data;
 using Conflux.Domain.Logic.Exceptions;
 using Conflux.Domain.Logic.Services;
-using Conflux.Domain.Models;
+using Conflux.Domain.Session;
 using Conflux.RepositoryConnections.SRAM;
 using Conflux.RepositoryConnections.SRAM.DTOs;
 using Microsoft.AspNetCore.Http;
