+++ Conflux.Domain.Logic/Services/ISessionMappingService.cs
@@ -3,7 +3,7 @@
 // 
 // © Copyright Utrecht University (Department of Information and Computing Sciences)
 
-using Conflux.Domain.Models;
+using Conflux.Domain.Session;
 
 namespace Conflux.Domain.Logic.Services;
