+++ Conflux.Domain.Logic/Services/ProjectsService.cs
@@ -6,7 +6,7 @@
 using Conflux.Data;
 using Conflux.Domain.Logic.DTOs;
 using Conflux.Domain.Logic.Exceptions;
-using Conflux.Domain.Models;
+using Conflux.Domain.Session;
 using Microsoft.EntityFrameworkCore;
 
 namespace Conflux.Domain.Logic.Services;
