+++ Conflux.Integrations.SRAM/CollaborationMapper.cs
@@ -5,7 +5,7 @@
 
 using Conflux.Data;
 using Conflux.Domain;
-using Conflux.Domain.Models;
+using Conflux.Domain.Session;
 using Conflux.RepositoryConnections.SRAM.DTOs;
 using Conflux.RepositoryConnections.SRAM.Exceptions;
 using Conflux.RepositoryConnections.SRAM.Models;
