+++ Conflux.Integrations.SRAM/ICollaborationMapper.cs
@@ -3,7 +3,7 @@
 // 
 // © Copyright Utrecht University (Department of Information and Computing Sciences)
 
-using Conflux.Domain.Models;
+using Conflux.Domain.Session;
 using Conflux.RepositoryConnections.SRAM.DTOs;
 
 namespace Conflux.RepositoryConnections.SRAM;
